make a `RancherOSConfig` class
the idea of the rancher instance class is to be able to make a `new RancherOSConfig` with chainable setters and getters. The class should also have a `.provisionOn(‘digitalocean’)` method signature. The idea is that if we ever want to `.provisionOn(‘aws’)` we could extend the class to do that.