fix type guard behavior
type guards (see: https://www.typescriptlang.org/docs/handbook/advanced-types.html ) work super if you aren’t comparing a type to ‘undefined’. In this case, we are. `authToken` can be either a string or undefined. It is undefined if `DIGITAL_OCEAN_PERSONAL_ACCESS_TOKEN` is not set in `.env`. Changing the type guard from a custom type guard to `typeof authToken …` (line 7) keeps the type guard from erroring out when the auth token is undefined.