poll for IP addresses
digital ocean doesn’t actually give a droplet an IP address until about 30 seconds after it is created. So, we have to keep polling until we get an IP address.

The polling loop continues until either the maximum number of seconds to poll has elapsed, or an ipv4 address has been found. If the droplet supports ipv6, the polling loop continues until either the maximum number of seconds to poll has elapsed, or both an ipv4 and ipv6 address are found.