set up ubuntu console, python on rancher instance
rancherOS ships with a busybox shell. That’s it. No ubuntu, no python, none of the stuff needed to run ansible scripts.

To change that, we need to give digital ocean a cloud-config directive.

The directive is in `src/provisionRancherOS/onDigitalOcean/cloud-config.yml`. It tells rancher to install the ubuntu console. This is a container that contains an ubuntu 18.04 image. Once the container is installed, set it as the default environment: i.e. make rancher mimic ubuntu. Finally, run `apt-get update` and `apt-get install -y python`. Once both are installed.

To send this cloud-config to digital ocean, we have to enable asset loading in webpack 5. To do that we add a `type: asset/resource` rule in `webpack.config.js` (see lines 20-23).

After that, we `require(“./cloud-config.yml”)` in `src/provisionRancherOS/onDigitalOcean/provisionOnDigitalOcean.ts` (line 12). This directs webpack’s asset loader to copy `src/provisionRancherOS/onDigitalOcean/cloud-config.yml` to `/dist`, assign a hash to it, and return the new path to the file to `cloudConfigPath`. Once webpack has done so, we point `fs.readFileSync(…)` to that path. It then reads the file and stores it in the variable `cloudConfig` (lines 12-17). Finally, we assign `cloudConfig` as the value for `user_data` (line 35), which in turn sends it off to digitalOcean.