make ansible playbook that configures rancher
run the ansible notebook with the following command

` ansible-playbook --extra-vars="rancher_ip=<ip of rancher droplet>” --inventory="./inventory.yml" install-owncloud-on-rancher.yml`

and runs the playbook `install-owncloud-on-rancher.yml` on the rancher instance created by `yarn serve`. Be sure to replace `<ip of rancher droplet>`, with you rancher instance’s IP address.

The playbook creates a new directory on the rancher instance: `/opt/owncloud`, and copies both `utils/owncloud-docker/.env` and `utils/owncloud-docker/docker-compose.yml` to it.

The playbook uses `inventory.yml` to determine how to connect to the rancher instance.