do not commit `package-lock.json` or `packages/**/yarn.lock` files
if `package-lock.json` shows up in a commit, it means that someone has run `npm install`. No one should be running `npm install` for any reason in this monorepo.

if `packages/**/yarn.lock` shows up in a commit, it means that someone has run `yarn` from within a package folder. No oneshould be running `yarn` for any reason.

to populate package.json, the correct command to run is:

`lerna bootstrap`