make sure that flags don’t clobber prompt answers
if a flag with the same name as a conditional prompt is passed, use the conditional prompt, rather than the flag.

the idea is that if you send a flag for a conditional prompt, and the conditional prompt isn’t shown, then the flag will be used. Else, the conditional prompt will be used - even if you set the flag.